requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
Jinja2==3.1.4
pydantic==2.9.2
httpx==0.27.2
